+++ project/app/Http/Controllers/UserController.php
@@ -13,6 +13,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $users = User::all();
         return view('users.index', compact('users'));
     }
@@ -22,6 +24,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         return view('users.create');
     }
 
@@ -30,6 +34,8 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'email' => 'required|email',
@@ -48,6 +54,8 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
 
         if (!$user->is_public && !auth()->check()) {
@@ -75,6 +83,8 @@ class UserController extends Controller
 
     public function follow(Request $request)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $userToFollow = User::findOrFail($request->user2_id);
         $loggedInUser = auth()->user();
 
@@ -97,6 +107,8 @@ class UserController extends Controller
 
     public function checkRequests()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = auth()->user();
         $pendingRequests = \App\Models\Follow::with('follower')
                                             ->where('user2_id', $user->id)
@@ -109,6 +121,8 @@ class UserController extends Controller
 
     public function accept($user1_id, $user2_id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $request = \App\Models\Follow::where('user1_id', $user1_id)
                                     ->where('user2_id', $user2_id)
                                     ->delete();
@@ -128,6 +142,8 @@ class UserController extends Controller
 
     public function reject($user1_id, $user2_id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $deleted = \App\Models\Follow::where('user1_id', $user1_id)
                                     ->where('user2_id', $user2_id)
                                     ->delete();
@@ -144,6 +160,8 @@ class UserController extends Controller
      */
     public function edit(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
 
         $this->authorize('update', $user);
@@ -155,6 +173,8 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         // Validate the incoming data
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
@@ -201,6 +221,8 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
         $user->delete();
 
@@ -208,34 +230,12 @@ class UserController extends Controller
     }
 
     /**
-     * Delete a user making it a anonymous user
-     */
-    public function deleteUser(string $id)
-    {
-        $user = User::findOrFail($id);
-        $user->username = null;
-        $user->firstname = null;
-        $user->surname = null;
-        $user->password = null;
-        $user->email = null;
-        $user->bio_description = null;
-        $user->profile_picture = 'default.jpg';
-        $user->is_public = false;
-        $user->type = 'deleted';
-        $user->save();
-
-        return redirect()->route('admin.users.manage')->with('success', 'User deleted successfully.');
-
-    }
-
-    
-
-
-    /**
      * Show the settings page.
      */
     public function settings()
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = auth()->user();
         return view('settings.show', compact('user'));
     }
@@ -245,6 +245,8 @@ class UserController extends Controller
      */
     public function updatePassword(Request $request, string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $validatedData = $request->validate([
             'old' => 'required|string|min:8',
             'new' => 'required|string|min:8',
@@ -266,6 +268,10 @@ class UserController extends Controller
      */
     public function deleteUser(string $id)
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
         $user->username = null;
         $user->firstname = null;
@@ -290,6 +296,8 @@ class UserController extends Controller
      */
     public function privacity(Request $request, string $id)   
     {
+        $loguser = auth()->user();
+        $this->authorize('banned', $loguser);
         $user = User::findOrFail($id);
 
         $validatedData = $request->validate([
